+++ proto/users/users_grpc.pb.go
@@ -17,11 +17,9 @@ const _ = grpc.SupportPackageIsVersion7
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type UsersClient interface {
-	GetUserByMail(ctx context.Context, in *GetByMailRequest, opts ...grpc.CallOption) (*GetUserResponse, error)
+	GetUserByMailOrT(ctx context.Context, in *GetByMailOrTRequest, opts ...grpc.CallOption) (*GetUserResponse, error)
 	GetUserByID(ctx context.Context, in *GetByIDRequest, opts ...grpc.CallOption) (*GetUserResponse, error)
 	FindUserByName(ctx context.Context, in *FindUserByNameRequest, opts ...grpc.CallOption) (*FindUserByNameResponse, error)
-	GetApproverInfo(ctx context.Context, in *GetApproverInfoRequest, opts ...grpc.CallOption) (*GetApproverInfoResponse, error)
-	CanApproveToUser(ctx context.Context, in *CanApproveToUserRequest, opts ...grpc.CallOption) (*CanApproveToUserResponse, error)
 }
 
 type usersClient struct {
@@ -32,9 +30,9 @@ func NewUsersClient(cc grpc.ClientConnInterface) UsersClient {
 	return &usersClient{cc}
 }
 
-func (c *usersClient) GetUserByMail(ctx context.Context, in *GetByMailRequest, opts ...grpc.CallOption) (*GetUserResponse, error) {
+func (c *usersClient) GetUserByMailOrT(ctx context.Context, in *GetByMailOrTRequest, opts ...grpc.CallOption) (*GetUserResponse, error) {
 	out := new(GetUserResponse)
-	err := c.cc.Invoke(ctx, "/users.Users/GetUserByMail", in, out, opts...)
+	err := c.cc.Invoke(ctx, "/users.Users/GetUserByMailOrT", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,33 +57,13 @@ func (c *usersClient) FindUserByName(ctx context.Context, in *FindUserByNameRequ
 	return out, nil
 }
 
-func (c *usersClient) GetApproverInfo(ctx context.Context, in *GetApproverInfoRequest, opts ...grpc.CallOption) (*GetApproverInfoResponse, error) {
-	out := new(GetApproverInfoResponse)
-	err := c.cc.Invoke(ctx, "/users.Users/GetApproverInfo", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *usersClient) CanApproveToUser(ctx context.Context, in *CanApproveToUserRequest, opts ...grpc.CallOption) (*CanApproveToUserResponse, error) {
-	out := new(CanApproveToUserResponse)
-	err := c.cc.Invoke(ctx, "/users.Users/CanApproveToUser", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
 // UsersServer is the server API for Users service.
 // All implementations must embed UnimplementedUsersServer
 // for forward compatibility
 type UsersServer interface {
-	GetUserByMail(context.Context, *GetByMailRequest) (*GetUserResponse, error)
+	GetUserByMailOrT(context.Context, *GetByMailOrTRequest) (*GetUserResponse, error)
 	GetUserByID(context.Context, *GetByIDRequest) (*GetUserResponse, error)
 	FindUserByName(context.Context, *FindUserByNameRequest) (*FindUserByNameResponse, error)
-	GetApproverInfo(context.Context, *GetApproverInfoRequest) (*GetApproverInfoResponse, error)
-	CanApproveToUser(context.Context, *CanApproveToUserRequest) (*CanApproveToUserResponse, error)
 	mustEmbedUnimplementedUsersServer()
 }
 
@@ -93,20 +71,14 @@ type UsersServer interface {
 type UnimplementedUsersServer struct {
 }
 
-func (UnimplementedUsersServer) GetUserByMail(context.Context, *GetByMailRequest) (*GetUserResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetUserByMail not implemented")
+func (UnimplementedUsersServer) GetUserByMailOrT(context.Context, *GetByMailOrTRequest) (*GetUserResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetUserByMailOrT not implemented")
 }
 func (UnimplementedUsersServer) GetUserByID(context.Context, *GetByIDRequest) (*GetUserResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetUserByID not implemented")
 }
 func (UnimplementedUsersServer) FindUserByName(context.Context, *FindUserByNameRequest) (*FindUserByNameResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method FindUserByName not implemented")
-}
-func (UnimplementedUsersServer) GetApproverInfo(context.Context, *GetApproverInfoRequest) (*GetApproverInfoResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetApproverInfo not implemented")
-}
-func (UnimplementedUsersServer) CanApproveToUser(context.Context, *CanApproveToUserRequest) (*CanApproveToUserResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method CanApproveToUser not implemented")
 }
 func (UnimplementedUsersServer) mustEmbedUnimplementedUsersServer() {}
 
@@ -121,20 +93,20 @@ func RegisterUsersServer(s grpc.ServiceRegistrar, srv UsersServer) {
 	s.RegisterService(&_Users_serviceDesc, srv)
 }
 
-func _Users_GetUserByMail_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(GetByMailRequest)
+func _Users_GetUserByMailOrT_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetByMailOrTRequest)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(UsersServer).GetUserByMail(ctx, in)
+		return srv.(UsersServer).GetUserByMailOrT(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/users.Users/GetUserByMail",
+		FullMethod: "/users.Users/GetUserByMailOrT",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(UsersServer).GetUserByMail(ctx, req.(*GetByMailRequest))
+		return srv.(UsersServer).GetUserByMailOrT(ctx, req.(*GetByMailOrTRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -175,49 +147,13 @@ func _Users_FindUserByName_Handler(srv interface{}, ctx context.Context, dec fun
 	return interceptor(ctx, in, info, handler)
 }
 
-func _Users_GetApproverInfo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(GetApproverInfoRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(UsersServer).GetApproverInfo(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/users.Users/GetApproverInfo",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(UsersServer).GetApproverInfo(ctx, req.(*GetApproverInfoRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Users_CanApproveToUser_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(CanApproveToUserRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(UsersServer).CanApproveToUser(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/users.Users/CanApproveToUser",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(UsersServer).CanApproveToUser(ctx, req.(*CanApproveToUserRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 var _Users_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "users.Users",
 	HandlerType: (*UsersServer)(nil),
 	Methods: []grpc.MethodDesc{
 		{
-			MethodName: "GetUserByMail",
-			Handler:    _Users_GetUserByMail_Handler,
+			MethodName: "GetUserByMailOrT",
+			Handler:    _Users_GetUserByMailOrT_Handler,
 		},
 		{
 			MethodName: "GetUserByID",
@@ -226,14 +162,6 @@ var _Users_serviceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "FindUserByName",
 			Handler:    _Users_FindUserByName_Handler,
-		},
-		{
-			MethodName: "GetApproverInfo",
-			Handler:    _Users_GetApproverInfo_Handler,
-		},
-		{
-			MethodName: "CanApproveToUser",
-			Handler:    _Users_CanApproveToUser_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},