+++ src/phonebook/phonebook.interface.ts
@@ -1,0 +1,10 @@
+export interface IPhoneBookUser {
+    _id?:string;
+    id: string;
+    armyId: string;
+    displayName: string;
+    firstName: string;
+    lastName: string;
+    rank: string;
+    mail: string;
+}